Move Argon2 formatting to argon2fmt module
