+++ src/lib/passhash/argon2fmt/argon2fmt.cpp
@@ -4,6 +4,7 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
+#include <botan/argon2fmt.h>
 #include <botan/argon2.h>
 #include <botan/rng.h>
 #include <botan/base64.h>
